requirements.txt
# Flask
# Flask-SQLAlchemy
# mysql-connector-python


Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.1
pymysql==1.1.0
